agent_langgraph/services/auth.py
from typing import Any, Dict, Optional
import jwt
from config import settings


def verify_token(token: str) -> Optional[Dict[str, Any]]:
    if not token:
        return None

    try:
        payload = jwt.decode(
            token,
            settings.JWT_SECRET_KEY,
            algorithms=[settings.JWT_ALGORITHM],
        )
        if not payload.get("userId"):
            return None
        return payload
    except jwt.ExpiredSignatureError:
        return None
    except jwt.InvalidTokenError:
        return None
