agent_langgraph/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.9.0
langgraph==0.2.43
python-dotenv==1.0.1
